requirements.txt
alt-profanity-check
openai
guardrails-ai
transformers
